Function detrojt::get_ty_const

pub fn get_ty_const<D: 'static>(key: usize) -> Option<D>

Get the data in the impl for the type that matches the given key. If the key is invalid, returns None.

Keys may be obtained using get_ty_const_key.

Due to limitations of the current implementation, calling this on an invalid key may sometimes cause arbitrary code execution (or a crash if you're lucky).

Implementation details

As noted in get_ty_const_key, the key is actually the memory address to the corresponding vtable. Extracting information from the vtable is pretty straightforward application of unsafe code.

The hard part is making sure the key is valid. Since we would rather return None than to segfault, the first steps is to ask the OS whether we can even read that memory. Then, we pull out the vtable and check whether it looks sensible.

Unfortunately, we have very little control over the contents of the vtable: most of it are just function pointers, and it's not easy to tell if a function pointer is right since every one is unique. The vtable does have size and alignment information, so we can in principle squeeze in a 128-bit magic number and check for that. Alas, this is limited by the fact that Rust doesn't like exabyte-sized arrays, and the fact that Rust has not yet implemented support for custom alignments (and even if it does, we don't know if it would support alignments that aren't powers of two). It would be even better if we could pick a random magic number for each build.