Trait detrojt::serde::Trait

pub trait Trait<S: InterSerialize, U: HasInterDeserialize> {
    fn serialize_inner(&self) -> Result<S, Box<Error>>;
    fn ty_const_key(&self) -> usize;
}

Supertrait for adding serialization support to user-defined traits.

If you want to define your own trait with serialization support, you will want to inherit from this trait and specify:

S: the intermediate serialization format,

U: the trait object type, which also indirectly specifies the intermediate deserialization format through HasInterDeserialize.

The intermediate serialization formats are used to work around the fact that trait objects have to be specialized for a concrete format. For the most part you can just pick a generic format such as serde_json::Value.

The trait object type U should be the boxed form of your user-defined trait. You probably have to wrap it in a newtype to avoid circular trait declarations.

In order for serialization to work, concrete implementations of your trait must satisfy the bound Into<U> + Serialize + DeserializeOwned + 'static.

Required Methods

fn serialize_inner(&self) -> Result<S, Box<Error>>

Serialize the inner object.

You should not implement this trait. It is automatically provided through the blanket implementation.

fn ty_const_key(&self) -> usize

Retrieve the key associated with the deserializer

You should not implement this trait. It is automatically provided through the blanket implementation.